Added resgiter page
