+++ src/fronted/Pages/Signup/Register.module.css
@@ -4,7 +4,7 @@
   align-items: center;
   margin-top: 1rem;
   width: 100%;
-  height: 80vh;
+  height: 100vh;
 }
 
 .signup_form {
